// src/core/domain/enums.ts
export enum PaymentStatus {
  CREATED = 'CREATED',
  PENDING = 'PENDING',
  REDIRECTED = 'REDIRECTED',
  CALLBACK_RECEIVED = 'CALLBACK_RECEIVED',
  SUCCESS = 'SUCCESS',
  FAILED = 'FAILED',
  CANCELLED = 'CANCELLED',
  REVERSED = 'REVERSED',
  REFUNDED = 'REFUNDED',
}

export enum TransactionType {
  PAYMENT = 'PAYMENT',
  VERIFY = 'VERIFY',
  REFUND = 'REFUND',
  REVERSE = 'REVERSE',
}

export enum GatewayCapability {
  CREATE_PAYMENT = 'CREATE_PAYMENT',
  VERIFY = 'VERIFY',
  INQUIRY = 'INQUIRY',
  REFUND = 'REFUND',
  REVERSE = 'REVERSE',
  WEBHOOK = 'WEBHOOK',
  TOKENIZATION = 'TOKENIZATION',
  RECURRING = 'RECURRING',
}